requirements.txt
lxml==5.1.0
httpx==0.27.0
trafilatura==1.9.0
pydantic==2.8.2
huggingface_hub>=0.26.0
jinja2==3.1.4
APScheduler==3.10.4
python-dotenv==1.0.1
supabase==2.4.6
loguru==0.7.2
fastapi==0.112.0
uvicorn==0.30.5
resend==2.0.0
PyYAML==6.0.2
pdfminer.six==20231228
